+++ fridgeApi/src/main/java/com/example/fridgeapi/services/UsersService.java
@@ -10,8 +10,8 @@ import java.util.List;
 public interface UsersService {
     public String createUser(Users user);
     public String updateUser(Users user);
-    public String deleteUser(BigInteger userId);
-    public Users getUser(BigInteger userId);
+    public String deleteUser(Long userId);
+    public Users getUser(Long userId);
     public List<Users> getAllUsers();
     public String loginUser(String email, String password);
 }
